rename http -> downloads
