+++ api-examples/pretrain-transformer-lm.py
@@ -14,7 +14,7 @@ from eight_mile.utils import Average, get_num_gpus_multiworker
 from eight_mile.pytorch.layers import init_distributed, save_checkpoint
 from eight_mile.pytorch.optz import *
 from baseline.pytorch.lm import TransformerLanguageModel, TransformerMaskedLanguageModel
-from eight_mile.http import DataDownloader
+from eight_mile.downloads import DataDownloader
 from transformer_utils import TensorWordDatasetReader, TensorCharDatasetReader, load_data_caching, get_lr_decay
 import numpy as np
 import codecs
